market_poc/requirements.txt
asgiref==3.8.1
Django==4.2.15
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
mssql-django==1.3
python-dotenv==1.0.0
djangorestframework==3.15.2
pandas==2.1.3
SQLAlchemy==2.0.23
PyJWT=2.10.1